+++ src/middlewares/index.js
@@ -1,5 +1,8 @@
 import useStorage from './useStorage';
 import useInstrumentation from './useInstrumentation';
 import useHttpLogger from './useHttpLogger';
+import traceRequest from './traceRequest';
 
-export { useStorage, useInstrumentation, useHttpLogger };
+export {
+  useStorage, useInstrumentation, useHttpLogger, traceRequest,
+};