+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `traceRequest` middleware.
+
 ## [0.2.0] - 16-03-2020
 
-## Added
+### Added
 
 - Add `useHttpLogger` middleware.